% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{app}
\alias{app}
\title{Launch Shiny App}
\usage{
app(...)
}
\arguments{
\item{...}{arguments to pass to shiny::runApp}
}
\description{
Launch a local copy of the shiny app that accompanies Simulating LMEM
}
