% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{appendix}
\alias{appendix}
\title{Open Simulating LMEM appendices}
\usage{
appendix(i = c("1a", "1b", "1c", "2", "3a", "3b"), filename = NULL)
}
\arguments{
\item{i}{Which appendix to open}

\item{filename}{Where to save the appendix}
}
\description{
Open Simulating LMEM appendices
}
