% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{paper}
\alias{paper}
\title{Open Simulating LMEM paper}
\usage{
paper(type = c("pdf", "html", "Rmd"))
}
\arguments{
\item{type}{The type of manuscript to open (pdf, html, or Rmd)}
}
\description{
Open the manuscript
}
